+++ models/bank_type.go
@@ -45,7 +45,7 @@ func (b *BankType) Delete() error {
 	return err
 }
 
-func (b *BankType) FindbyID(id int) error {
+func (b *BankType) FindbyID(id uint64) error {
 	err := basemodel.FindbyID(&b, id)
 	return err
 }
